add new product to default products from adminpage
