+++ src/Components/Products/ProductItem/ProductItem.tsx
@@ -9,7 +9,7 @@ import { Link } from "react-router-dom";
 
 const ProductItem = (props: { price: number; id: string; name: string; brand: string; image: string; description: string; amount: number }) => {
     const cartCtx = useContext(CartContext);
-    const price = `${props.price.toFixed(2)}:-`;
+    const price = `${props.price},00:-`;
 
     const addToCartHandler = (amount: number) => {
         cartCtx.addItem({
